requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2.post1
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.4.2
